Fix the use of open_flags during FUSE_OPEN
The flags passed as an argument to FUSE_OPEN and those returned in its
reply are distinct.  The former are things like `O_RDWR`, where the
latter are things like `FOPEN_DIRECT_IO`.  So don't copy the former to
the latter.  Also, always set `FOPEN_KEEP_CACHE`, since this is a
read-only file system.