Fix size-only lsextattr and getextattr
If the user supplies size=0 for getextattr or lsextattr, that means to
lookup the size of the attribute's value or the list's length,
respectively, without fetching it.

Fix a logic inversion in the way we calculate this size.  Before, by
luck it would return something too large.  So userspace would
successfully get the value, but would have to allocate far too much
memory.