Add tests for file systems with 1kB blocksize
Using a second golden image with a 1kB blocksize has two purposes:

* It finds bugs in how we handle images with non-default block size or
  inode size.  Two such minor bugs are fixed in this commit.

* It allows us to test BTrees that would be impractically large if they
  used the larger block size and directory size.  For instance, creating
  a BTree of height 3 on the old golden image would've required > 2M
  files.  In the new image, it only requires 128k.

Also, add a SUPERBLOCK global variable.  It's unfortunate that we must
do this, because it prevents libxfuse from working with two disk images
at a time, but we need to know some information from the superblock
within bincode::de::Decode::decode impls.  So it's either user the
global variable, or else stop using bincode.

Finally, raise MSRV to 1.70.0.