Remove most sanity functions
The only reason that sanity() existed was because it needs SuperBlock,
and Decode::decode doesn't have access to that.  But since we now have a
global SUPERBLOCK anyway, we may as well inline the sanity checking into
the Decode impls.

But leave in DinodeCore::sanity, because that struct is so big it would
be annoying to manually implement Decode.