Fix fsblock translation with non-power-of-two disk sizes
The way that XFS maps file system block numbers to disk offsets is
complicated. Throughout most of our code, we've been using the simple
formula fsblock * sb_blocksize. That only works when
sb_agblocks & ((1 << sb_agblklog) - 1) == 0. This condition seems to be
true if the file system size is a power of two, which it as in all of
our test cases. But it isn't generally true.

The symptom of this problem is that when mounting an image where the
file system size is not a power of two, various panics happen caused by
reading the wrong disk block. Usually the panic messages are something
like "UUID mismatch" or "magic mismatch".

Fixes #93