Fix reading data from some BTree-formatted files
In a BTree-formatted file, the precise location of the bmbt pointers is
not documented.  It seems from experiment that the location always lies
at a byte address divisible by 8.  This matters for some files that have
xattrs.  The precise length of the xattr probably matters.

I don't know, for lack of real-world examples, whether this same
adjustment is necessary for directories.  But I assume it is.