Add FUSE_NO_OPEN_SUPPORT and FUSE_NO_OPENDIR_SUPPORT
If the kernel supports it (and recent versions of FreeBSD do), it will
now skip calling FUSE_OPEN and FUSE_OPENDIR, saving a little bit of
time.

Also, set FUSE_EXPORT_SUPPORT, which will allow the mountpoint to be
reexported over NFS.

Finally, for completeness set FUSE_ASYNC_READ, though this is a no-op
because fuser does it by default.