Add test case for reading a BTree-formatted file with a hole at the end
Such a test case already existed for extent-list formatted files.  Also,
fix a bug in FileBtree::get_extent: it would report the wrong length for
a hole that extended to EoF.  The bug was harmless.  I don't know of any
way that it would be observable to a user.  The likeliest way that it
would be observable would be if an OS bug caused the OS to attempt to
read past EoF.