Fix reading inodes that aren't sectorsize-aligned
XFS V4 can use 256 byte inodes, which are 256 byte-aligned on disk.  But
most disks require at least reads to be at least 512B aligned.  This is
not a problem that can be fixed by BufReader alone.  Fix it by importing
the BlockReader code from Benjamin Stürz's fuse-ufs project.  That
ensures that all reads will be both sectorsize-aligned and a multiple of
the sectorsize.

Enhanced BlockReader by using DIOCGSECTORSIZE, if possible, and by
allowing the internal buffer to be extended when desirable.